cs8803/requirements.txt
antlr4-python3-runtime==4.8
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
graphviz==0.8.4
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
jedi==0.16.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.2.0
mccabe==0.6.1
more-itertools==8.2.0
mxnet==1.6.0
mypy==0.761
mypy-extensions==0.4.3
numpy==1.18.1
onnx==1.6.0
orderedset==2.0.3
packaging==20.3
parso==0.6.2
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.11.3
psutil==5.7.0
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
xgboost==1.0.2
zipp==3.1.0
